+++ projects/client/Unit/RabbitMQ.Client.Unit.csproj
@@ -95,6 +95,7 @@
     <Compile Include="src\unit\TestConnectionRecovery.cs" />
     <Compile Include="src\unit\TestConnectionShutdown.cs" />
     <Compile Include="src\unit\TestConnectionWithBackgroundThreads.cs" />
+    <Compile Include="src\unit\TestInitialConnection.cs" />
     <Compile Include="src\unit\TestConsumerCancelNotify.cs" />
     <Compile Include="src\unit\TestConsumerCount.cs" />
     <Compile Include="src\unit\TestConsumerExceptions.cs" />
